Format check and test 1990/dds/README.md
In order to try and speed things along some of the things that might be
good to have in code blocks (as in '`') have not been done so it can be
seen if this can be omitted. If it works out it's almost certain that
this can be done with other README.md files which also have places where
'`'s might be useful. An example here where it's not used but might be
is in the section 'Program commands' with variables NOT enclosed in '`'s
and other things that might look a little better with them.