Updates to 1990/tbr - alt code
Properly document the fact that there is alt code. This involved a note
about the fixes in 2023 as the alt code had to have fixes applied too
and it was referred to in the description of it (that was not referred
to as alt code).

I have not yet checked the README.md file for formatting or typo issues.