Typo/format check 1990/westley/README.md
This also fixes a typo I introduced that would result in a syntax error
in use.

It also adds proper '### Alternate build' and '### Alternate use'
sections.

I've decided that for the rest of the README.md files I will try not
worrying much about grammar unless it makes it harder to understand the
information or it's very quick. This way all these can get done sooner.