Add alt code to 2020/tsoj
This version will feel more at home for those who are familiar with vi
movement keys. It does not use curses so no arrow keys but the entry
itself uses the old wasd movement controls. Now one can use k, h, j and
l for movement.

Updated .gitignore and README.md file. The file was NOT format fixed.
Work on this entry was not supposed to happen yet but looking for
something I came across this one and decided to make it more friendly to
vi users (the wasd movement has always been a problem for me another
winner in 2020 also told me that they found it difficult).