Update bug status of 1996/jonth + format README.md
There are two issues to be aware of, both features, that have been added
to the bugs.md. The status has been added to the README.md as well.
These issues are that if X is not running the program is likely to crash
and also the two windows are likely to be on top of each other so
you'll have to move one off the other.

Since it's been requested and since I am editing the README.md file
anyway the award and author information has been removed from the file
so that the first line is '## How to build:'. Other README.md files are
also currently modified and this will happen that way before a mass
edit.