Minor fix in 1998/dlowe/README.md wrt alt code
The description said to run 'make dlowe.alt' and this is incorrect: it
should be 'make alt'.

Also since it has been requested that the award and author information
be removed from the README.md files this has been done since this fix
was already made but not yet committed.

In the previous commit it was stated that there were five more README.md
files counting this one, but that's incorrect as that one only was
removing the home(?) address of one of the authors (two I believe) so
that can happen with the mass edit of removing award and author
information. Thus not counting this one there are another three
README.md files I believe it is before that mass edit can happen.