Fix 1991/rince and add two alt versions
The program was supposed to print a message that indicates whether you
won or lost but it did not show up because of endwin() being called
after it. To keep the code as close to the original as possible that
part was left in but an additional call to printf() was made after
endwin() is called.

The first alt version removes the maximum number of moves so that
everyone can finish, even people not good at this kind of puzzle. The
second alt version allows you to reconfigure (at compile time) the
maximum number of moves even if you wish to make it harder to win.

Format fix README.md file.

At this time I believe the edits necessary for 1991/rince are complete.