Drop 1991/fine back to its original 80 chars
With more clever use of the Makefile the number of characters is back
down to 80 which is what the original was!

This also fixes the warning that was introduced when it was dropped back
down to 85. 'B' is defined as '(int)b' which is used in two places where
a cast is necessary or otherwise will silent a warning.

The README.md file has been checked and this should complete this entry
now.