Updates and fixes to 1991/davidguy
The README.md file has been format and typo checked.

The Makefile was updated slightly to disable implicit int warning and
also to include certain header files for the syscalls used. This is not
strictly necessary but since some systems (like macOS) do have problems
with some functions not being declared this was done as it does not have
an adverse effect.