Updates to 1991/brnstnd
Since the author refers to the tool as 'sorta' it is now a symlink to
the program. Updated .gitignore.

The sorta.README.md file has been format fixed.

The README.md has been partially checked. The reason for this commit
when it's not is I need to find where a problem in formatting is and in
order to do that I need to see the rendered html (a code block is not
terminated it would appear).