Add 1992/westley/try.sh + format fixes
The try.sh script runs the program for the world cities and the alt
version for the US cities, one at a time, that the author suggested, as
well as the approximate location that the entry was judged, noted from
the judges.

The README.md has been format fixed.

There is a possible bug in the alt version which is in the bugs.md file.
There is also a feature in the bugs.md which does introduce a strange
look in the README.md file:

    STATUS: INABIAF - please **DO NOT** fix
    STATUS: known bug - please help us fix

but they are for different things.