Fix 1992/albert, swap alt/main code
The alt code was the original; the entry was a bug fixed version by Leo
(before he was a judge). However that version, the fixed version, was
not obfuscated and as the judges say in the FAQ:

    the contest is called the IOCCC, not the INVOCCC :-)

so it seems better to have the version that has a bug as the entry. But
that code no longer compiled. This has been fixed.

Updated .gitignore.

Minor changes to README.md wrt the swapping of alt code/entry code. The
formatting has not been checked and no typos have been fixed either.
These will come later.