Fixes to 1992/ant
The Makefile has had a new rule added and another fixed as 'am' does not
support certain features that the Makefile uses.

Added #include <time.h> as the author stated that for some systems it
might be necessary (like DOS with Turbo C assuming anyone here uses
that).

Typo fixes including references to files that do not exist (am.c ->
ant.c and so on).

Typo/format check README.md and ant.README.