Format/typo check 1993/vanb + add alt code
The alt code was provided by the author as a kind of spoiler but I fixed
it to also work with clang like I did with the original entry.

Silenced some warnings (all on macOS are now silenced, at least with
default cc which is clang).

Format/typo check README.md file.

This should complete 1993/vanb.