Some fixes in 1993/rince
It was stated, perhaps even by me but no longer sure of this (it looks
like my style) that after the program is terminated the terminal might
not be sane and one could do 'stty sane', 'reset' and other things.
Instead I have changed the code to just call endwin() and removed that
information instead. This way the user should not be burdened by it. The
alt code that I had added earlier on also has this fix.

The markdown files have not been completely checked but both README.md
and design.md have had some fixes applied.

The thanks file also points out that Yusuke's fix applies the suggested
fix for select(2) found in the README.md.