Add another alt version to 1993/rince + formatting
The formatting of the README.md file has been completed. Typos were also
fixed.

The formatting of the design.md has been completed. Typos were also
fixed.

The second alt version makes movement for the default game (which is in
the source code itself) more familiar to vi users. It is an exercise for
others to change the somewhat awkward keys for other games. Looking at
design.md might be of help there.

The bugs.md has been updated with a set of features, not bugs, that
should no be fixed.