Fixes in 1993/leo
Running 'make' now links 'mind' to 'leo' (via the 'mind' rule).

Format/typo check the README.md file.

This should complete the entry unless a certain thing I found is a bug
in which case the bugs.md has to be updated (and the README.md refer to
it). This is not certain yet.