Updates to 1993/cmills - alt code + bugs + format
The README.md file has been format/typo checked.

The alt version uses usleep(3) with a configurable time (at compile
time). This version is suggested first as it apparently goes too fast in
modern systems. I say apparently as it doesn't appear to work with
macOS. I don't know if it works in other platforms but it is in bugs.md,
that there is a problem.