Minor fix in 1993/cmills/Makefile
When adding the alt code I yank pasted the rule for cmills but by
accident neglected to add the '.alt' to the rule. This was not seen
because of the way the Makefile 'alt' rule is designed (in some cases
one need not have more than just that 'alt' rule) and other reasons
besides.