Fixes in 1994/dodsond2 + add alt code
Fixed bug where in some cases the program would enter an infinite loop,
forcing one to kill the program.

Fixed bug where only if you had more than one arrow would shooting it
take you to the room you shoot into (it now will do it even if you're
using the last arrow). This fix is done the same way as the above fix
was made as they both have to do with shooting the final arrow.

The alt code allows one to cheat by configuring (at compile time) the
number of arrows to start out with. If one loses (as in stolen) all
their arrows it will be (of course) 0 but if one also gets to the robber
they'll get back the 3 + starting number of arrows (in the original this
is just 3) assuming they have had that many stolen. That is a
(mis)feature or maybe a bug that is carried over from the original entry
that might or might not want to be fixed. I described how it might be
done in the bugs.md but as I don't have time to do it now, focusing
instead on the README.md files of the remaining entries.  The alt
version might allow one to more easily test this.

Besides the above (mis)feature or bug added to the bugs.md some features
that should not be fixed were added to bugs.md. It is questionable
whether or not the above should be fixed but it seems like since there
is dead code, if one ends up shooting an arrow then they cannot get
back the arrows that were stolen even they have have lost all that they
did not shoot, that it should be fixed. But it might be argued that
you're not supposed to get any back if you shoot them since in that case
these were already spent/used. At the same time one could argue that you
should at least get back the ones that were stolen.