Minor fix in 1994/dodsond2 alt code
The original code checks for equality of number of arrows stolen before
giving back the arrows. If the number is not reset after recovering
arrows is set back to 0 that means that you can only recover them once.
The alt code checked for >= though and now it's been changed to be
equality like the original code.