Further improvements to 1991/dds fix
Rather than having the string that is run via system(3) use
-Wno-return-type the generated code (also in the string) now returns 1;
rather than 'return;' so that main() has a proper return statement.

This means the string was updated and the call to system(3) was also
updated but it is a bit better now. It is hoped that this is the last
time (..yet again..) that this will have to be updated to work with
various systems but as before the magic is in the thanks file, at least
for now.