Add 1996/gandalf/try.sh and format README.md
The script is rather elaborate in that it runs the programs in the
different ways the judges previously suggested as well as some
additional runs and shows the difference (or if no difference says they
are identical) of the different commands, using temporary files that are
deleted if the script is sent (some) signals, the script is quit (when
prompted to press a key pressing 'q' will quit the script) or if the
script runs to completion.

The README.md file was format/typo checked.

BTW: it is absolutely perilous to try the patience of Gandalf
(https://www.glyphweb.com/arda/g/gandalf.html). Go ahead, try it! :-)

The Makefile has had -Wno-implicit-int added to CSILENCE.