Add try.sh and try.alt.sh to 1996/westley
These scripts simplify running the three other scripts (that the author
provided, renamed to have the extension '.sh') to show the three styles
of clocks with both the fixed (for modern systems) version and the
original version (the alt code).

The README.md file has been format and typo checked.

This should complete 1996/westley.