Add alt build to 1998/banks + format/typo checks
The alt build uses the same code but sets a default for those who don't
have page up and page down keys (like Macs). The file keysym.h that I
added has some of the keys (macro name and comment indicating what
character it corresponds to) from X11 (same lines in fact).

The alt code/build/use sections explain the defaults and how to
reconfigure it. The author provided more information on these details
and one can certainly do it with the main build too. The use of this alt
build is it sets a default for those without the page up and page down
but still allowing one to configure it differently. The keys were
somewhat arbitrarily selected to be on the left side of a QWERTY
keyboard as using the arrow keys and the right side of the keyboard at
the same time might be harder :-).

The README.md file has been typo/format checked and fixed and this
should complete 1998/banks I do believe.