Fixes and other updates to 1998/schweikh3
The constant that was in the code was changed to a macro in the Makefile
called SZ. Though it's highly unlikely that this will ever be needed the
author noted that increasing the value has some merit in a rare case so
it's now easier to do this (though one might have to type more,
admittedly, as all the -D options have to be specified but it no longer
is required to modify the code).

The README.md file has been format and typo checked. I did change the
way the man page was displayed (in the README.md) so this has to be
verified to look good but everything else should be good anyway.

The try.sh script I added was updated to have another command that was
in the judges' remarks. Instead it's now all in the try section and the
judges' remarks just have the comments, not the command to try.

When I updated the thanks file I noticed a couple things that could be
clarified in another entry so those minor fixes were made too but have
no relation to this commit.