Update 1998/dlowe - add scripts + formatting
The following scripts were added:

- try.sh: a fun use of the program.
- pootify.sh: a local pootifier that downloads a web page (or use a
local file) and either changes it to text or writes html which can be
viewed in a local browser.
- pootify.cgi.sh: a CGI version that the judges added but made to use
bash(1) and pass shellcheck(1). There is an issue that is known with
this and the 'Historical remarks' discusses this in more detail as well
as how to use the pootify.sh and pootify.cgi.sh scripts.

The README.md file has been formatted as well.