Fixes in 1998/df
A pointer to int was being used for fopen(3) and this has been changed
to a FILE * as using the wrong type has been observed in some cases to
cause programs to fail.

Make the code more like the original by using the old style main() arg
so that it could read 'int main(i,love_unix)' and also changing the
typedef for 'int' from 'lint' to be '_int' (see the code for why it no
longer can be 'int' and ask yourself why it can be 'int' in other cases).

Format/typo check README.md file.