Update 1988/westley/try.sh
Allow one to specify CC=foo to compile with a different compiler. If CC
is empty it will be set to 'cc'. One can specify a path in which case
it'll be that path; otherwise the shell will work it out. Of course if
it's not a compiler that's another matter entirely.

Also remove check for 'westley' existing already because 'make' will do
that via the Makefile.

And although not strictly necessary the ints used in the program were
changed to floats as they are floats.

Updated CSILENCE in Makefile.