Add 2000/dhyang/try.sh, make void main int main
The try.sh script is as usual: replaces what the judges recommended one
try.

It was suggested to me a while back that when I encounter a void main
that I change it to int to be more portable so I have done that.

README.md file typo/format checked.

This should complete the review of 2000/dhyang.