Updates to 2001/kev
The code and alt code now have '#ifndef..#endif' blocks so that one need
not specify all the -D options if they wish to change just one. In the
case of SPEED if it's undefined it is set to 50 which is what the
Makefile originally had it as but I changed it to 100 as it seemed to be
a problem. The author does note how different speeds make for different
kinds of play and so this change is even more useful.

Format/typo check README.md file.

This should complete the review of 2001/kev and I believe only one entry
remains now.