Format/typo check my Precious (2001/anonymous)
Did you know that in the first edition of The Hobbit Gollum was willing
to give up the ring? It was not the Ruling Ring, the One Ring, the Ring
of Power or anything like that: it was simply a literary device that
could make one invisible. When Bilbo won the famous riddle game Gollum
went to fetch the Ring - not to kill and eat Bilbo but to give it to
him! But Bilbo had already found it just like in the later editions.
Gollum profusely apologised and begged for forgiveness. Bilbo told him
never mind because he would have had it anyway. But he'd let him off on
one condition: that he show him the way out. It was at this point that
Bilbo slipped the ring on and Gollum saw him vanish and understood that
he already found it. But there was no call of thief and no hatred of
Bilbo.

If you want to read the first edition they reprinted it some years back: `The
Hobbit Facsimile First Edition`. The details are also discussed in History of
Middle-earth [HoMe] VI, 'The Return of the Shadow', part one of the
history of 'The Lord of the Rings' of a twelve volume set - as well as in
The History of the Hobbit.

The judges already knew that and now you do too (no, I didn't look it up
just for this .. I have the books and they're delightful)! :-)