Add alt versions to 2004/gavin + format/typo check
There already was one alt version which I found from the author's
website about the entry; they used the version for development and it's
unobfuscated. But two new ones were added.

The most important feature is that it allows one to change the image
size along with the anti-alias setting at compile time. This is
gavare.alt.c. The second version (which also has the size/anti-alias
changes) should theoretically work for Windows. This is gavare.alt2.c
and running make alt will not try and compile it as most it'll fail on
sane systems. Instead try make alt2. Both of these versions have some
form of protection with the size/anti-alias setting and one does not
have to define all three to get it to compile due to
'#ifndef..#elif..#endif' in the code.

The README.md file has been format/typo fixed.