Fix and improve 2005/giljade/try.sh
After one exited less(1) the text about leaving the file there for one
to look at it with an editor was on the same line as other output.

Also use less -X so that the screen is not cleared. That way one can see
the last part of the file that they looked at.