Add 2005/sykes/try.sh and test.sh + make test rule
The try.sh is different from others in that it is a loop showing a menu
with different options to use the emulator. It notes at the start of
each iteration that one has to hit ctrl-c or whatever they have set
their interrupt as.

The test.sh is run by make test. The exit status is ignored as it
returns non-zero.

Format/typo check README.md.

I am not done with this entry as the bugs.md should probably be updated
to note some things the author noted. This will have to come another
time.