Add how to find intr/etc. to FAQ
Since some entries require intr/EOF etc. I have added to the FAQ how to
find it. This was inspired by the 2005/sykes as to end the program one
has to send intr. For those with grep -o it's possible to show just the
output; other ways exist but those ways are more complicated and I've
not gone there.