Slight clarification in 1985/lycklama
I believe that entries that alt versions are recommended to try first
should say what it looks like in modern systems so much as what it
looked like with systems of the time it was published.

Other entries with this have to be located as well so I've put it in the
todo.md file too.