Add input.txt to 2005/boutines + format/typo check
The input.txt file is what the author included in their remarks as
input to try out. This has been integrated into one of the commands to
try in the try section (which did not even have a command to try out -
it just referred to the command in the to use section).

BTW: the author mentioned the Mersenne prime. Did you know that Landon
Curt Noll has held eight world records related to large primes?