Updates to 2006/sykes1
Add alt code based on author's remarks.

Add try.sh along the lines of most the others.

Some of the README.md has been formatted. More has to be done another
time.

Add the video file of the cube being solved in case that video is also
lost (like the previous one was).

The pdf file I had added was incomplete. For some reason it was only one
page, not the full document. This has been fixed.

When I can do the README.md file that should complete the entry though
it's possible other issues will be found at that point.