Add 2006/birken/try.sh, bugs.md, format/typo check
The try.sh runs the program on all the tofu input files, prompting the
user to press a key before running each one so they have time to look at
the output of the previous one.

The README.md file format and typo checked.

This should complete 2006/birken except that it uses gets(), not
fgets(), which has been added to the bugs.md file. I almost got it to
work but it crashes with computer.tofu so I did not change it. Will look
at this later, maybe, but the bugs.md has a diff of the single change
that almost works. I did not have the time to write where it fails but I
do happen to know at least the end result.