Minor improvements in thanks file
Moved 'README and thanks' to top and made it '.. and thanks for all the
fish. :-)' instead.. This is because it talked about how the judges
wish to thank the various people but at the bottom it would serve no
purpose and yet it felt like it should still be in the file. The heading
or subheading title is an experiment but I think it is probably better
than 'README and thanks'.