Add second arg to main() in 1985/august
Due to clang defect in number of args allowed in main(). Some versions
will if you use, say, 4 args, object and report that main() can only
have 0, 2 or 3 args. That same version does not appear to object to 1
arg (bug?) but it's entirely possible that this changes so I have
changed the number of args of main() to 2 even though it's unused just
out of abundance of caution.

Updated CSILENCE for no implicit int. I have to test the unused
parameter warning. That might need to be added to,
-Wno-unused-parameter, but I have to find out which compilers will or
will not object to it and that can come later on.