Add alt versions to 2011/toledo + formatting
The first version allows one to reconfigure the controls and the size of
the game field (which previously was done by changing the Makefile and
updating the code respectively but now is possible at compilation time)
and the second version uses the author's supplied layer.c file that
should work for Windows.

The keysym.h file is actually keysymdef.h from X11 so that users can see
the macro names of the keys.

The README.md file has been format/typo checked.

This should complete the review of 2011/toledo.