Add try.sh and hint.md to 2012/hou + formatting
The hint.md file is necessary as the '#' header in it specifies the
title and stylesheet as part of the entry.

The try.sh script is as usual.

Format/typo check README.md file. The hint.md file formatting was not
changed due to the regexp engine amongst other reasons. Just obvious
spelling errors were updated.