Add 2012/tromp/try.sh, update bugs.md + formatting
The try.sh has a variety of commands, some that come from the judges and
some that come from the author.

The bugs.md has features added, not bugs.

The README.md file has been format and typo checked.

The how.md file has been format and typo checked.

The author stated that there is a macro that indicates the term space
but as it's a big number and I don't see it being a problem I did not
(unlike in other cases) add alt code that lets one redefine it.