improve references to people in thanks-for-fixes.md
Added sub-headings to the "Thank you honor roll"
section for each person credited for a fix.

Sorted the sub-headings in the "Thank you honor roll" section.

Added a final "Did we neglect to credit you?", in case
we forget to credit someone's fix later on, requesting
corrections to the thank you record.

For each IOCCC winning entry section, the first mention of the person
fixing the entry becomes a link down to their sub-headings in the
"Thank you honor roll" section.  Subsequent mentions of the
same person in the same IOCCC winning entry section are **NOT**
turned into links: only the 1st reference is linked.

Removed trailing whitespace.

Fixed word case for Linux.

Added reference links for the first mention of ShellCheck, SDL, etc.
in a given IOCCC winning entry section.