Add alt versions to 2012/endoh1 + formatting
The alt versions let one control how long to sleep between writes as in
modern systems it goes pretty fast. There are two, one for the colour
version that was added by Yusuke at the request of the judges and one
for the original. But the one for the original is called endoh1.alt2.c
as Yusuke provided endoh1.alt.c as a deobfuscated version.

The new alt versions let one only redefine certain macros should they
choose but they can define them all or not define them all because of
the #ifndef..#endif guards. These macros include the sleep time (S) as
well as the others that Yusuke suggested one might want to change.

The README.md file has been format/typo checked.